app/requirements.txt
aioconsole==0.8.1
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiokafka==0.12.0
aiomonitor==0.7.1
aiosignal==1.4.0
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.6.4
avro==1.12.0
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
ciso8601==2.3.3
click==8.1.8
colorlog==6.9.0
confluent-kafka==2.11.1
croniter==2.0.7
cryptography==46.0.1
Cython==3.1.4
exceptiongroup==1.3.0
fastavro==1.12.0
faust-cchardet==2.1.19
faust-streaming==0.11.3
faust-streaming-rocksdb==0.9.3
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
intervaltree==3.1.0
janus==2.0.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mode-streaming==0.4.1
multidict==6.6.4
mypy_extensions==1.1.0
opentracing==2.4.0
orjson==3.11.3
packaging==25.0
prompt_toolkit==3.0.52
propcache==0.4.0
pycares==4.11.0
pycparser==2.23
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
setproctitle==1.3.7
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
telnetlib3==2.0.8
terminaltables==3.1.10
trafaret==2.1.1
typing_extensions==4.15.0
urllib3==2.5.0
venusian==3.1.0
wcwidth==0.2.14
yarl==1.20.1
